+++ src/main/util/include/foreign_library.h
@@ -11,4 +11,5 @@ double add_two_numbers(double, double);
 double sub_two_numbers(double, double);
 double mult_two_numbers(double, double);
 double *div_two_numbers(double *, double);
-unsigned long int random_longint();
+unsigned long long int random_longint();
+
